micro_login/requirements.txt
asgiref==3.8.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fonttools==4.55.4
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
